+++ medivh-core/src/main/kotlin/tech/medivh/core/Medivh.kt
@@ -7,13 +7,10 @@ import net.bytebuddy.matcher.ElementMatchers
 import tech.medivh.api.DebugTime
 import tech.medivh.core.env.MedivhContext
 import tech.medivh.core.jfr.JfrInterceptor
-import java.io.File
-import java.io.FileInputStream
-import java.io.FileOutputStream
+import tech.medivh.core.reporter.TagMethod
+import tech.medivh.core.reporter.TagMethodLogFileWriter
 import java.lang.instrument.Instrumentation
 import java.util.*
-import java.util.zip.ZipEntry
-import java.util.zip.ZipInputStream
 
 
 /**
@@ -23,28 +20,12 @@ object Medivh {
 
     @JvmStatic
     fun premain(agentArgs: String?, inst: Instrumentation) {
-
         val properties = String(Base64.getUrlDecoder().decode(agentArgs))
         val context = MedivhContext(properties)
-
+        context.reportDir().mkdirs()
+        val writer = TagMethodLogFileWriter(context.reportDir().resolve(TagMethod.FILE_NAME))
         Runtime.getRuntime().addShutdownHook(Thread {
-            val timeReport = context.mode().timeReport
-            // dir = /build/medivh/reports/time/uuid
-            val dir = context.reportDir()
-            val reportZip = dir.parentFile.parentFile.parentFile.resolve("medivh-report.zip")
-            val reportDir = dir.resolve("report/")
-            reportDir.mkdirs()
-            unzip(reportZip, reportDir)
-            MedivhJsGenerator(context).generateJs()
-            reportDir.resolve(timeReport.htmlTemplateName()).copyTo(reportDir.resolve("index.html.temp")).apply {
-                this.parentFile.listFiles { file -> file.extension == "html" }?.forEach {
-                    check(it.delete()) { i18n(context.language(), "error.fileDeleteFailed", it.path) }
-                }
-                val indexHtml = reportDir.resolve("index.html")
-                check(this.renameTo(indexHtml)) { i18n(context.language(), "error.fileRenameFailed", indexHtml.path) }
-                //  in windows, the path is like "C:\Users\xxx" but idea console can't recognize it is a file path
-                println(i18n(context.language(), "tip.seeReport", indexHtml.absolutePath.replace("\\", "/")))
-            }
+            writer.close()
         })
 
         val mode = context.mode()
@@ -60,10 +41,10 @@ object Medivh {
                     debugTime?.let {
                         val debugTimeDesc = resolveDebugTime(it)
                         mode.timeReport.setup(MethodSetup(MethodToken.fromMethodDescription(method), debugTimeDesc))
+                        writer.writeMethod(TagMethod(method.name, desc.name, debugTimeDesc.expectTime))
                     }
                 }
-                // builder.method(ElementMatchers.isAnnotatedWith(DebugTime::class.java))
-                builder.method(ElementMatchers.any())
+                builder.method(ElementMatchers.isAnnotatedWith(DebugTime::class.java))
                     .intercept(advice)
             }.installOn(inst)
 
@@ -74,26 +55,5 @@ object Medivh {
         return DebugTimeDesc(debugTime.getValue("expectTime").resolve() as Long)
     }
 
-    private fun unzip(zipFile: File, reportDir: File) {
-        if (reportDir.resolve("report/js").exists()) {
-            return
-        }
-        ZipInputStream(FileInputStream(zipFile)).use { zis ->
-            var entry: ZipEntry? = zis.nextEntry
-            while (entry != null) {
-                val newFile = File(reportDir, entry.name)
-                if (entry.isDirectory) {
-                    newFile.mkdirs()
-                } else {
-                    newFile.parentFile.mkdirs()
-                    FileOutputStream(newFile).use { fos ->
-                        zis.copyTo(fos)
-                    }
-                }
-                zis.closeEntry()
-                entry = zis.nextEntry
-            }
-        }
-    }
 }