+++ medivh-core/src/main/kotlin/tech/medivh/core/statistic/MultiThreadStatistic.kt
@@ -42,4 +42,4 @@ open class MultiThreadStatistic(val methodName: String) {
     }
 }
 
-data class ThreadInvokeInfo(val threadName: String, val invokeInfo: InvokeInfo)
+