+++ medivh-core/src/main/kotlin/tech/medivh/core/statistic/ThreadInvokeInfo.kt
@@ -1,0 +1,8 @@
+package tech.medivh.core.statistic
+
+import tech.medivh.core.InvokeInfo
+
+/**
+ * @author gongxuanzhangmelt@gmail.com
+ */
+data class ThreadInvokeInfo(val threadName: String, val invokeInfo: InvokeInfo)