+++ doc/CHANGELOG.md
@@ -1,2 +1,3 @@
 0.2.3
- - support auto language detection
+
+- support auto language detection